+++ src/lib/field.unit-test.ts
@@ -7,7 +7,16 @@ import { Provable } from './provable.js';
 import { Binable } from '../bindings/lib/binable.js';
 import { ProvableExtended } from './circuit_value.js';
 import { FieldType } from './field.js';
-import { createEquivalenceTesters, throwError } from './testing/equivalent.js';
+import {
+  equivalentProvable as equivalent,
+  oneOf,
+  field,
+  bigintField,
+  throwError,
+  unit,
+  bool,
+  Spec,
+} from './testing/equivalent.js';
 
 // types
 Field satisfies Provable<Field>;
@@ -56,73 +65,75 @@ test(Random.field, Random.int(-5, 5), (x, k) => {
   deepEqual(Field(x + BigInt(k) * Field.ORDER), Field(x));
 });
 
+// Field | bigint parameter
+let fieldOrBigint = oneOf(field, bigintField);
+
 // special generator
 let SmallField = Random.reject(
   Random.field,
   (x) => x.toString(2).length > Fp.sizeInBits - 2
 );
-
-let { equivalent1, equivalent2, equivalentVoid1, equivalentVoid2 } =
-  createEquivalenceTesters(Field, Field);
+let smallField: Spec<bigint, Field> = { ...field, rng: SmallField };
+let smallBigint: Spec<bigint, bigint> = { ...bigintField, rng: SmallField };
+let smallFieldOrBigint = oneOf(smallField, smallBigint);
 
 // arithmetic, both in- and outside provable code
-equivalent2((x, y) => x.add(y), Fp.add);
-equivalent1((x) => x.neg(), Fp.negate);
-equivalent2((x, y) => x.sub(y), Fp.sub);
-equivalent2((x, y) => x.mul(y), Fp.mul);
+let equivalent1 = equivalent({ from: [field], to: field });
+let equivalent2 = equivalent({ from: [field, fieldOrBigint], to: field });
+
+equivalent2(Fp.add, (x, y) => x.add(y));
+equivalent1(Fp.negate, (x) => x.neg());
+equivalent2(Fp.sub, (x, y) => x.sub(y));
+equivalent2(Fp.mul, (x, y) => x.mul(y));
 equivalent1(
-  (x) => x.inv(),
-  (x) => Fp.inverse(x) ?? throwError('division by 0')
+  (x) => Fp.inverse(x) ?? throwError('division by 0'),
+  (x) => x.inv()
 );
 equivalent2(
-  (x, y) => x.div(y),
-  (x, y) => Fp.div(x, y) ?? throwError('division by 0')
+  (x, y) => Fp.div(x, y) ?? throwError('division by 0'),
+  (x, y) => x.div(y)
 );
-equivalent1((x) => x.square(), Fp.square);
+equivalent1(Fp.square, (x) => x.square());
 equivalent1(
-  (x) => x.sqrt(),
-  (x) => Fp.sqrt(x) ?? throwError('no sqrt')
+  (x) => Fp.sqrt(x) ?? throwError('no sqrt'),
+  (x) => x.sqrt()
 );
-equivalent2(
-  (x, y) => x.equals(y).toField(),
-  (x, y) => BigInt(x === y)
+equivalent({ from: [field, fieldOrBigint], to: bool })(
+  (x, y) => x === y,
+  (x, y) => x.equals(y)
 );
-equivalent2(
-  (x, y) => x.lessThan(y).toField(),
-  (x, y) => BigInt(x < y),
-  SmallField
+
+equivalent({ from: [smallField, smallFieldOrBigint], to: bool })(
+  (x, y) => x < y,
+  (x, y) => x.lessThan(y)
 );
-equivalent2(
-  (x, y) => x.lessThanOrEqual(y).toField(),
-  (x, y) => BigInt(x <= y),
-  SmallField
+equivalent({ from: [smallField, smallFieldOrBigint], to: bool })(
+  (x, y) => x <= y,
+  (x, y) => x.lessThanOrEqual(y)
 );
-equivalentVoid2(
-  (x, y) => x.assertEquals(y),
-  (x, y) => x === y || throwError('not equal')
+equivalent({ from: [field, fieldOrBigint], to: unit })(
+  (x, y) => x === y || throwError('not equal'),
+  (x, y) => x.assertEquals(y)
 );
-equivalentVoid2(
-  (x, y) => x.assertNotEquals(y),
-  (x, y) => x !== y || throwError('equal')
+equivalent({ from: [field, fieldOrBigint], to: unit })(
+  (x, y) => x !== y || throwError('equal'),
+  (x, y) => x.assertNotEquals(y)
 );
-equivalentVoid2(
-  (x, y) => x.assertLessThan(y),
+equivalent({ from: [smallField, smallFieldOrBigint], to: unit })(
   (x, y) => x < y || throwError('not less than'),
-  SmallField
+  (x, y) => x.assertLessThan(y)
 );
-equivalentVoid2(
-  (x, y) => x.assertLessThanOrEqual(y),
+equivalent({ from: [smallField, smallFieldOrBigint], to: unit })(
   (x, y) => x <= y || throwError('not less than or equal'),
-  SmallField
+  (x, y) => x.assertLessThanOrEqual(y)
 );
-equivalentVoid1(
-  (x) => x.assertBool(),
-  (x) => x === 0n || x === 1n || throwError('not boolean')
+equivalent({ from: [field], to: unit })(
+  (x) => x === 0n || x === 1n || throwError('not boolean'),
+  (x) => x.assertBool()
 );
-equivalent1(
-  (x) => x.isEven().toField(),
-  (x) => BigInt((x & 1n) === 0n),
-  SmallField
+equivalent({ from: [smallField], to: bool })(
+  (x) => (x & 1n) === 0n,
+  (x) => x.isEven()
 );
 
 // non-constant field vars